[py-osacomponent] now uses release version of appscript 0.19.x
git-svn-id: https://appscript.svn.sourceforge.net/svnroot/appscript@657 5e899d65-cff2-4c64-b955-3fa930e4d869

+++ py-osacomponent/trunk/src/PyOSA/main.c
@@ -54,7 +54,7 @@ ComponentResult PyOSA_main(ComponentParameters *params, Handle ciStorage) {
 		err = badComponentSelector;
 	}
 	#ifdef DEBUG_ON
-	if (err) fprintf(stderr, "PyOSA_main returned error: %i\n", err);
+	if (err) fprintf(stderr, "PyOSA_main returned error: %i\n", (int)err);
 	#endif
 	return err;
 }
